requirements-dev.txt
# requirements-dev.txt
black
bandit
python-dotenv
django-storages
pip-tools
pylint
pylint-django
-r ./requirements.txt